Changed Event structure to work with actions.
Changed Events.jsx to display Event components. Event components data is
maged by EventStore and EventActions
