+++ src/stores/EventStore.jsx
@@ -1,0 +1,14 @@
+import { EventEmitter } from 'events';
+
+class EventStore extends EventEmitter {
+  constructor() {
+    super();
+    this.event = [
+      {},
+      {},
+    ];
+  }
+}
+
+const eventStore = new EventStore;
+export default eventStore;
